Used in Ubuntu 18.04 to manipulate desktops as atomic elements of an array, e.g., reordering and inserting/deleting empties.
An addition for examples of `xdotool` usage.   This is a working example I have used in parallel with the Ubuntu 18.04 / Gnome WM to make managing desktops much easier.  By binding these function to keys the following fundtions are provided::  

 - move the desktops up or down in the ordered list, 
 - insert empty desktops anywhere,
 - purge all the empty desktops 

Can be used in conjunction with Gnome's bird's eye view of the desktops (Super+a, Super+a, move mouse to right edge).
Requires Gnome "dynamic/static desktop" option to be set to static -  because this program performs it's own dynamic.

Because `xdotool` does not depend on Gnome, this program could probably work with other WM's using a linear array of desktops.